Moved webxr.d.ts from extras/types to renderers/webxr
